Move channel implementation to a dedicated file
